+++ src/core/server/saved_objects/saved_objects_service.mock.ts
@@ -79,6 +79,7 @@ const createSetupContractMock = () => {
     addClientWrapper: jest.fn(),
     registerType: jest.fn(),
     getImportExportObjectLimit: jest.fn(),
+    setRepositoryFactoryProvider: jest.fn(),
   };
 
   setupContract.getImportExportObjectLimit.mockReturnValue(100);