+++ src/plugins/saved_objects_management/public/management_section/objects_table/saved_objects_table.test.tsx
@@ -161,6 +161,7 @@ describe('SavedObjectsTable', () => {
       goInspectObject: () => {},
       canGoInApp: () => true,
       search,
+      dateFormat: 'YYYY-MM-DD',
     };
 
     findObjectsMock.mockImplementation(() => ({